backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
djoser==2.2.3
flake8==7.1.0
isort==5.13.2
django-filter==24.2
Pillow==10.4.0
gunicorn==20.1.0
python-dotenv==1.0.1
psycopg2==2.9.9